+++ tests/tests/src/nested_records_with_type_params.mjs
@@ -3,6 +3,9 @@
 
 let options = {
   extra: {
+    extraField: {
+      theField: true
+    },
     name: "test",
     superExtra: {
       age: 2222
@@ -18,6 +21,9 @@ let options = {
 
 let opts2 = {
   extra: {
+    extraField: {
+      theField: true
+    },
     name: "test",
     superExtra: {
       age: "1234"